requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
livereload==2.7.1
MarkupSafe==3.0.2
openai==1.60.2
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
watchdog==6.0.0
Werkzeug==3.1.3
WTForms==3.1.2
email-validator==2.1.1
psycopg2-binary==2.9.9
